refactor: split runner spike command facade
